Improve file saving
- Check if file exists before downloading
- Integrate properly with tqdm

+++ utils.py
@@ -195,20 +195,28 @@ def get_attachment_data(attachment_url: str):
 
     return response.content
 
-def save_attachment(attachment_url: str):
-    local_endpoint = attachment_url.removeprefix("https://trello.com/1/").replace("/", "-")
-    file_location = os.path.join(os.getenv("ATTACHMENT_DIRECTORY"), local_endpoint)
+def save_attachment(attachment_url: str, tqdm=None):
+    local_file_name = attachment_url.removeprefix("https://trello.com/1/").replace("/", "-")
+    file_location = os.path.join(os.getenv("ATTACHMENT_DIRECTORY"), local_file_name)
+    local_endpoint = os.path.join(os.getenv('ATTACHMENT_SERVER_ENDPOINT'), local_file_name)
 
     # Ensure the attachments directory exists.
     os.makedirs(os.getenv("ATTACHMENT_DIRECTORY"), exist_ok=True)
+
+    # If file already exists, skip download.
+    if os.path.exists(file_location):
+        if tqdm is not None:
+            tqdm.write(f"\tAttachment already exists at {file_location}, skipping download.")
+        return local_endpoint
     
     image_data = get_attachment_data(attachment_url)
 
     with open(file_location, "wb") as f:
         f.write(image_data)
-    print(f"Saved attachment to {file_location}")
+    if tqdm is not None:
+        tqdm.write(f"\tSaved attachment to {file_location}")
 
-    return os.path.join(os.getenv('ATTACHMENT_SERVER_ENDPOINT'), local_endpoint)
+    return local_endpoint
 
 def get_time_from_id(card_id: str):
     return datetime.fromtimestamp(int(card_id[0:8], 16))
